Move obstacles to new models folder
