+++ src/dev/btietz/multiTerrain/AppMultiTerrain.h
@@ -34,7 +34,7 @@
 #include "dev/btietz/kinematicString/KinematicSpineCPGControl.h"
 
 // obstacles
-#include "dev/btietz/Obstacles/tgBlockField.h"
+#include "models/obstacles/tgBlockField.h"
 
 // This library
 #include "core/tgModel.h"
