+++ src/models/obstacles/tgBlockField.cpp
