+++ src/models/obstacles/tgStairs.cpp
